Added log statement
log("value is %i\n", value);

Now works. Scanner adds support to strings, but they are only used in the log statement. the Log op has the number of args in the lower 4 bits, followed by the length of the format string, followed by the string itself. Works on both Mac and Arduino.

The main reason for doing this is to write tests that will output their results. It costs about 2K in code, but it will be worth it when the system is testable