Changed SetFrame to have 8 bit locals
One of the test cases had 16 local vars, which failed. 16 is fine for passed params, but not really for locals, especially in this case where there is a 10 element array and a few other vars. So now SetFrame puts the params in the lower 4 bits and the next byte is the locals. Takes up the same amount of space.