Added the ability to put a var statement anywhere in the code.
SetFrame needs to be at the top of the function and it contains the amount of local storage needed. So I write it as 0 at the top of the function, remember that location and then update it at the end when I know the amount of storage needed. Changed the tests to put var statements where they are used and it all works