Got rid of foreach, added for and makde if and else extended opcodes.
I had to make the If and Else opcodes extended so all opcodes doing a jump have 12 bit relative addresses. That makes resolving the jump addresses easier.

Now foreach is gone and there is a proper for loop. All modes are supported. You can have an initializer that either is declared in the for loop or outside, with proper block scoping (not tested yet)

There are several things I don't like about this implementation. First, it's very general, so there are a couple of jumps that aren't needed if you have a simple case where some of the for loop contents are missing. Second, there are way too many jumps because I am generating code inline. So the iteration, which doesn't happen until the end of the loop is generated at the beginning, so you have to jump around it and then jump back to do the iteration. I plan on doing deferred generation so I can get rid of the extra jumps. But for now it works correctly.